Add Redis as application backend

+++ config-service/src/main/java/com/alexlloyd/configservice/service/ConfigServiceImpl.java
@@ -1,6 +1,7 @@
 package com.alexlloyd.configservice.service;
 
 import java.util.Collection;
+import java.util.Map;
 
 import com.alexlloyd.configservice.api.ConfigDAO;
 import com.alexlloyd.configservice.api.ConfigService;
@@ -34,6 +35,14 @@ public class ConfigServiceImpl implements ConfigService {
      */
     @Override
     public void createConfig(String configName) throws ConfigAlreadyExistsException {
+        if (configName == null) {
+            throw new NullPointerException("Config name is null");
+        }
+
+        if (this.configDAO.hasConfig(configName)) {
+            throw new ConfigAlreadyExistsException(configName);
+        }
+
         this.configDAO.createConfig(configName);
     }
 
@@ -46,7 +55,13 @@ public class ConfigServiceImpl implements ConfigService {
      */
     @Override
     public Config getConfig(String configName) throws ConfigDoesNotExistException {
-        return this.configDAO.getConfig(configName);
+        if (!this.configDAO.hasConfig(configName)) {
+            throw new ConfigDoesNotExistException(configName);
+        }
+
+        Map<String, String> configMap = this.configDAO.getConfigMap(configName);
+
+        return new Config(configMap);
     }
 
     /**
@@ -59,8 +74,10 @@ public class ConfigServiceImpl implements ConfigService {
      */
     @Override
     public void updateConfig(String configName, String key, String value) throws ConfigDoesNotExistException {
-        Config config = this.getConfig(configName);
-        config.addConfig(key, value);
+        if (!this.configDAO.hasConfig(configName)) {
+            throw new ConfigDoesNotExistException(configName);
+        }
+        this.configDAO.updateConfig(configName, key, value);
     }
 
     /**
@@ -71,6 +88,9 @@ public class ConfigServiceImpl implements ConfigService {
      */
     @Override
     public void deleteConfig(String configName) throws ConfigDoesNotExistException {
+        if (!this.configDAO.hasConfig(configName)) {
+            throw new ConfigDoesNotExistException(configName);
+        }
         this.configDAO.deleteConfig(configName);
     }
 
@@ -83,7 +103,11 @@ public class ConfigServiceImpl implements ConfigService {
      */
     @Override
     public void deleteValue(String configName, String key) throws ConfigDoesNotExistException {
-        this.getConfig(configName).deleteKey(key);
+        if (!this.configDAO.hasConfig(configName)) {
+            throw new ConfigDoesNotExistException(configName);
+        }
+
+        this.configDAO.deleteValue(configName, key);
     }
 
     /**
@@ -93,6 +117,6 @@ public class ConfigServiceImpl implements ConfigService {
      */
     @Override
     public Collection<String> getConfigNames() {
-        return this.configDAO.listConfigs().keySet();
+        return this.configDAO.listConfigs();
     }
 }
